fix(nvim): disable Treesitter for Haskell files
This (temporary) fix disables Treesitter for Haskell files. This is
necessitated by the fact that TS causes nvim to crash when opening my
XMonad configuration file. This allows for the editing (and cleanup) of
XMonad configuration within nvim but should be removed after XMonad
configuration has been cleaned up and so that TS can be made available
when editing Haskell files.